requirements.txt
asgiref==3.7.2
attrs==23.2.0
backports.zoneinfo==0.2.1
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
importlib-resources==6.1.1
inflection==0.5.1
isort==5.13.2
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
pkgutil-resolve-name==1.3.10
psycopg==3.1.17
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
regex==2023.12.25
rpds-py==0.17.1
sqlparse==0.4.4
typing-extensions==4.9.0
uritemplate==4.1.1
zipp==3.17.0
python-decouple==3.5
